ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.20.3",
    "keras>=3.6.0",
    "manim-ml>=0.0.24",
    "manim>=0.18.1",
    "matplotlib>=3.9.2",
    "numpy>=1.26.4",
    "pydot>=3.0.2",
    "pygame>=2.6.1",
    "ruff>=0.7.3",
    "seaborn>=0.13.2",
    "stubs>=1.0.0",
    "tensorflow>=2.18.0",
    "torchvision>=0.20.1",
    "torchviz>=0.0.2",
    "tqdm>=4.67.0",
    "imageio>=2.36.0",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.7.3",
]
